Change component name to directory
