+++ src/components/directory/directory.styles.scss
@@ -1,4 +1,4 @@
-.categories-container {
+.directory-container {
   width: 100%;
   display: flex;
   flex-wrap: wrap;
